# Reduce build times of PersistenceDiagram modules