--- a/paraview/xmls/PersistenceDiagramApproximation.xml
+++ b/paraview/xmls/PersistenceDiagramApproximation.xml
@@ -43,15 +43,14 @@
           <Group name="filters"/>
         </ProxyGroupDomain>
         <DataTypeDomain name="input_type">
-          <DataType value="vtkDataSet"/>
+          <DataType value="vtkImageData"/>
         </DataTypeDomain>
         <InputArrayDomain name="input_scalars" number_of_components="1">
           <Property name="Input" function="FieldDataSelection" />
         </InputArrayDomain>
         <Documentation>
           Data-set to process.
-          Input should be a regular grid (.vti), otherwise FTM will be used
-          to compute the persistence diagram, and other fields will be empty.
+          Input should be a regular grid (.vti).
         </Documentation>
       </InputProperty>