This short PR reduces the build times of the `ttkPersistenceDiagram`, `ttkPersistenceDiagramApproximation` and `ttkTrackingFromFields` modules through these two optimizations:

- the computation of saddle-saddle pairs with the FTM backend has been deleted, since it is now superseded by the DiscreteMorseSandwich algorithm.
- `PersistenceDiagramApproximation` now calls only `executeApproximateTopology` in the base class instead of the way heavier `execute`. The filter has been made to be executed only on regular grids in order to get rid of its fallback to FTM.

The observed reduction in build times is as follow (on my computer with some template instantiations disabled):
* `ttkPersistenceDiagram`: from 180 to 110s,
* `ttkTrackingFromFields`: from 185 to 110s,
* `ttkPersistenceDiagramApproximation`: from 180 to 30s.

Enjoy,
Pierre